requirements.txt
# Core
playwright>=1.48.0
html2text>=2024.2.26
pydantic>=2.10.0
python-dotenv>=1.0.0

# LLM Providers
google-generativeai>=0.8.0
openai>=1.50.0
